+++ src/latextopretext.ts
@@ -19,6 +19,7 @@ function convertMath(text: string) {
 
   return result;
 }
+
 function convertTextMarkup(text: string) {
   //Convert verbatim
   let result = text.replace(
@@ -111,82 +112,6 @@ function makeLines(text: string) {
   // console.log("LINES: "+lines );
   // console.log("LINES LENGTH: "+lines.length );
   return lines;
-  // let loopCount = text.split(/\r\n|\r|\n/).length;
-  // let depth = 0;
-  // let block = "";
-  // console.log("TEXT ARRAY: " + lines);
-  // console.log("LOOP COUNT: " + lines.length);
-  // let i = 0;
-  // while (i < lines.length) {
-  //   console.log("LINE: " + lines[i] + "; BLOCK: " + block);
-  //   // If we are at depth 0 and we see regular text, assume we are at a start of a paragraph.  Read until an empty line or a \begin is encountered.
-  //   if (depth === 0) {
-  //     if (lines[i].startsWith("\\begin")) {
-  //       depth += 1;
-  //       i += 1;
-  //     } else if (lines[i].startsWith("\\end")) {
-  //       lines[i-1] = lines[i-1]+"\n"+lines[i].
-  //       depth -= 1;
-  //     }
-
-  //   }
-  //   if (convertCheck) {
-  //     if (lines[i].trim().length === 0) {
-  //       block += "\n";
-  //       blockList.push(block);
-  //       block = "";
-  //       continue;
-  //     }
-  //     if (lines[i].trim().startsWith("\\begin")) {
-  //       block += lines[i].trim() + "\n";
-
-  //       convertCheck = false;
-  //     } else {
-  //       if (i === 0) {
-  //         block += lines[i].trim() + "\n";
-  //         if (lines[i + 1].trim().length === 0) {
-  //           blockList.push(block);
-  //           block = "";
-  //         }
-  //       } else if (i + 1 === loopCount) {
-  //         block += lines[i].trim() + "\n";
-  //         blockList.push(block);
-  //         block = "";
-  //       } else {
-  //         block += converter(lines[i]).trim() + "\n";
-  //         if (lines[i + 1].trim().length === 0) {
-  //           blockList.push(block);
-  //           block = "";
-  //         }
-  //       }
-  //     }
-  //   } else {
-  //     if (lines[i].trim().startsWith("\\end")) {
-  //       block += lines[i].trim();
-
-  //       blockList.push(block);
-  //       block = "";
-
-  //       convertCheck = true;
-  //     } else {
-  //       if (lines[i].trim().length === 0) {
-  //         block += "\n";
-  //       } else {
-  //         block += lines[i].trim() + "\n";
-  //       }
-  //       if (i + 1 === loopCount) {
-  //         blockList.push(block);
-  //       }
-  //     }
-  //   }
-  // }
-
-  // //log each block in terminal to see how it is building as test
-  // for (var blockCheck of blockList) {
-  //   console.log("BLOCK CHECK: " + blockCheck);
-  // }
-
-  // return blockList;
 }
 
 function convertLines(lines: string[]) {
@@ -231,17 +156,7 @@ function convertLines(lines: string[]) {
   return result;
 }
 
-// Converts full text testing convert line by line//calls each conversion in extension
-// export function latexToPretext(text: string) {
-//   var result = text;
 
-//   //result = convertParagraph(result);
-//   result = convertMath(result);
-//   result = convertTextMarkup(result);
-//   result = convertQuotation(result);
-
-//   return result;
-// }
 export function latexToPretext(text: string) {
   var result = convertLines(makeLines(text));