+++ CHANGELOG.md
@@ -9,7 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- C
 - New snippets/completions
+  
 
 ## [0.17.0] - 2024-02-25