+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [UNRELEASED]
 
+## [] - 2024-07-06
+
 ### Added
 
 - Blank Lines formatting option to specify whether no, some, or many blank lines are added between elements.